+++ tests/files/src/0727_generic_array.php
@@ -24,6 +24,9 @@ class Base {
         return [$this->item, $this->item];
     }
 
+    /**
+     * @return T
+     */
     public function getOne() {
         return $this->item;
     }