+++ tests/files/src/0872_noop_closure.php
@@ -4,6 +4,6 @@ function test_str_replace($value)  {
         $value = preg_replace_callback('/ab/', function ($_) {
             return '';
         }, $value, -1, $count);
-    } while ($count);
+    } while ($count);  // TODO: Figure out reason for PhanPossiblyInfiniteLoop
     return $value;
 }